+++ packages/account-tree-controller/src/types.ts
@@ -23,10 +23,8 @@ import type {
 } from '@metamask/profile-sync-controller';
 import type { GetSnap as SnapControllerGetSnap } from '@metamask/snaps-controllers';
 
-import type {
-  AccountTreeController,
-  controllerName,
-} from './AccountTreeController';
+import type { controllerName } from './AccountTreeController';
+import type { AccountTreeControllerMethodActions } from './AccountTreeController-method-action-types';
 import type {
   BackupAndSyncAnalyticsEventPayload,
   BackupAndSyncEmitAnalyticsEventParams,
@@ -79,46 +77,6 @@ export type AccountTreeControllerGetStateAction = ControllerGetStateAction<
   AccountTreeControllerState
 >;
 
-export type AccountTreeControllerSetSelectedAccountGroupAction = {
-  type: `${typeof controllerName}:setSelectedAccountGroup`;
-  handler: AccountTreeController['setSelectedAccountGroup'];
-};
-
-export type AccountTreeControllerGetSelectedAccountGroupAction = {
-  type: `${typeof controllerName}:getSelectedAccountGroup`;
-  handler: AccountTreeController['getSelectedAccountGroup'];
-};
-
-export type AccountTreeControllerGetAccountsFromSelectedAccountGroupAction = {
-  type: `${typeof controllerName}:getAccountsFromSelectedAccountGroup`;
-  handler: AccountTreeController['getAccountsFromSelectedAccountGroup'];
-};
-
-export type AccountTreeControllerGetAccountContextAction = {
-  type: `${typeof controllerName}:getAccountContext`;
-  handler: AccountTreeController['getAccountContext'];
-};
-
-export type AccountTreeControllerSetAccountWalletNameAction = {
-  type: `${typeof controllerName}:setAccountWalletName`;
-  handler: AccountTreeController['setAccountWalletName'];
-};
-
-export type AccountTreeControllerSetAccountGroupNameAction = {
-  type: `${typeof controllerName}:setAccountGroupName`;
-  handler: AccountTreeController['setAccountGroupName'];
-};
-
-export type AccountTreeControllerSetAccountGroupHiddenAction = {
-  type: `${typeof controllerName}:setAccountGroupHidden`;
-  handler: AccountTreeController['setAccountGroupHidden'];
-};
-
-export type AccountTreeControllerSetAccountGroupPinnedAction = {
-  type: `${typeof controllerName}:setAccountGroupPinned`;
-  handler: AccountTreeController['setAccountGroupPinned'];
-};
-
 export type AllowedActions =
   | AccountsControllerGetAccountAction
   | AccountsControllerGetSelectedMultichainAccountAction
@@ -127,23 +85,16 @@ export type AllowedActions =
   | KeyringControllerGetStateAction
   | SnapControllerGetSnap
   | UserStorageController.UserStorageControllerGetStateAction
-  | UserStorageController.UserStorageControllerPerformGetStorage
-  | UserStorageController.UserStorageControllerPerformGetStorageAllFeatureEntries
-  | UserStorageController.UserStorageControllerPerformSetStorage
-  | UserStorageController.UserStorageControllerPerformBatchSetStorage
-  | AuthenticationController.AuthenticationControllerGetSessionProfile
+  | UserStorageController.UserStorageControllerPerformGetStorageAction
+  | UserStorageController.UserStorageControllerPerformGetStorageAllFeatureEntriesAction
+  | UserStorageController.UserStorageControllerPerformSetStorageAction
+  | UserStorageController.UserStorageControllerPerformBatchSetStorageAction
+  | AuthenticationController.AuthenticationControllerGetSessionProfileAction
   | MultichainAccountServiceCreateMultichainAccountGroupAction;
 
 export type AccountTreeControllerActions =
   | AccountTreeControllerGetStateAction
-  | AccountTreeControllerSetSelectedAccountGroupAction
-  | AccountTreeControllerGetSelectedAccountGroupAction
-  | AccountTreeControllerGetAccountsFromSelectedAccountGroupAction
-  | AccountTreeControllerGetAccountContextAction
-  | AccountTreeControllerSetAccountWalletNameAction
-  | AccountTreeControllerSetAccountGroupNameAction
-  | AccountTreeControllerSetAccountGroupPinnedAction
-  | AccountTreeControllerSetAccountGroupHiddenAction;
+  | AccountTreeControllerMethodActions;
 
 export type AccountTreeControllerStateChangeEvent = ControllerStateChangeEvent<
   typeof controllerName,